requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy
mysql-connector-python
python-jose[cryptography]
pydantic[email]
passlib[bcrypt]
python-multipart